+++ src/test/kotlin/net/pwall/json/schema/JSONSchemaNonstandardFormatTest.kt
@@ -1,8 +1,8 @@
 /*
- * @(#) JSONSchemaCustomValidatorTest.kt
+ * @(#) JSONSchemaNonstandardFormatTest.kt
  *
  * json-kotlin-schema Kotlin implementation of JSON Schema
- * Copyright (c) 2020 Peter Wall
+ * Copyright (c) 2021 Peter Wall
  *
  * Permission is hereby granted, free of charge, to any person obtaining a copy
  * of this software and associated documentation files (the "Software"), to deal
@@ -28,27 +28,22 @@ package net.pwall.json.schema
 import kotlin.test.Test
 import kotlin.test.expect
 import kotlin.test.fail
+
 import net.pwall.json.JSON
-import net.pwall.json.JSONString
 import net.pwall.json.schema.parser.Parser
 import net.pwall.json.schema.validation.StringValidator
 
-class JSONSchemaCustomValidatorTest {
+class JSONSchemaNonstandardFormatTest {
 
-    @Test fun `should make use of custom validator`() {
+    @Test fun `should make use of nonstandard format replacement validation`() {
         val parser = Parser()
-        parser.customValidationHandler = { key, uri, location, value ->
-            when (key) {
-                "x-test" -> {
-                    if (value is JSONString && value.get() == "not-empty")
-                        StringValidator(uri, location, StringValidator.ValidationType.MIN_LENGTH, 1)
-                    else
-                        fail("Unknown type")
-                }
+        parser.nonstandardFormatHandler = { keyword, uri, location ->
+            when (keyword) {
+                "not-empty" -> StringValidator(uri, location, StringValidator.ValidationType.MIN_LENGTH, 1)
                 else -> null
             }
         }
-        val filename = "src/test/resources/test-custom-validator.schema.json"
+        val filename = "src/test/resources/test-nonstandard-format.schema.json"
         val schema = parser.parseFile(filename)
         val json1 = JSON.parse("""{"aaa":"Q"}""")
         expect(true) { schema.validate(json1) }
@@ -61,19 +56,19 @@ class JSONSchemaCustomValidatorTest {
         expect(3) { errors.size }
         errors[0].let {
             expect("#") { it.keywordLocation }
-            expect("http://pwall.net/test-custom#") { it.absoluteKeywordLocation }
+            expect("http://pwall.net/test-nonstandard-format#") { it.absoluteKeywordLocation }
             expect("#") { it.instanceLocation }
             expect("A subschema had errors") { it.error }
         }
         errors[1].let {
             expect("#/properties/aaa") { it.keywordLocation }
-            expect("http://pwall.net/test-custom#/properties/aaa") { it.absoluteKeywordLocation }
+            expect("http://pwall.net/test-nonstandard-format#/properties/aaa") { it.absoluteKeywordLocation }
             expect("#/aaa") { it.instanceLocation }
             expect("A subschema had errors") { it.error }
         }
         errors[2].let {
-            expect("#/properties/aaa/x-test") { it.keywordLocation }
-            expect("http://pwall.net/test-custom#/properties/aaa/x-test") { it.absoluteKeywordLocation }
+            expect("#/properties/aaa/format") { it.keywordLocation }
+            expect("http://pwall.net/test-nonstandard-format#/properties/aaa/format") { it.absoluteKeywordLocation }
             expect("#/aaa") { it.instanceLocation }
             expect("String fails length check: minLength 1, was 0") { it.error }
         }